+++ src/about.css
@@ -209,20 +209,26 @@
     padding:20px !important;    
     cursor: pointer;
   }
-  .navabt, .navhome, .navstates{
+  .navabt, .navhome{
     color:white !important;
     background-color: grey;
     margin:0 20px !important;
     width:10vw !important;
     text-align: center;
   }
-  
+  .navstates{
+    color:white !important;
+    background-color: grey;
+    margin:0 20px !important;
+    width:22vw !important;
+    text-align: center;
+  }
   
   .dropdown-toggle{
        color:white !important;
   }
   .logoleft{
-    background-image: url(./images/titlerr.png);
+    background-image: url(./states/bihar/images/titlerr.png);
     width:245px;
     height:50px;
     background-repeat: no-repeat;
@@ -230,7 +236,7 @@
   .logo{
     width:140px;
     height:40px;
-    background-image: url(./images/logo2.png);
+    background-image: url(./states/bihar/images/logo2.png);
 
   }
   .App-footer {
@@ -278,7 +284,7 @@
 }
 .bgimage{
   position:absolute !important;
-  background-image: url(./images/bg2.png);
+  background-image: url(./states/bihar/images/bg2.png);
   background-repeat: no-repeat;
   width:27vw;
   height:55vh;
@@ -396,7 +402,7 @@
   background-color:#001b5b;
   top:290%;
   position:absolute;
-  background-image: url(./images/below7.png);
+  background-image: url(./states/bihar/images/below7.png);
   background-repeat: no-repeat;
   left:-8%;
 }
