+++ src/states/bihar/child.js
@@ -1,5 +1,5 @@
 import Table from 'react-bootstrap/Table';
-import "../App.css";
+import "../../App.css";
 export default function ChildComponent(props){
 
     const arr = ["...","KATIHAR", "MOTIHARI", "MUZAFFARPUR", "MUNGER", "NALANDA", "PATNA", "ROHTAS", "SUPAUL"];
