+++ src/states/bihar/images/lft.png
