+++ src/states/bihar/images/threebars.png
