+++ src/states/bihar/imgs.js
