+++ src/states/bihar/slideshow/MovieCard/index.js
