+++ src/states/bihar/slideshow/SimpleSlider/slick-theme.css
