+++ src/states/bihar/slideshow/SimpleSlider/slick.css
