+++ src/states/bihar/updatebox.js
