+++ src/tables/courses/courseBihar.js
@@ -1,9 +1,8 @@
-import TextTruncate from "./truncate";
+import TextTruncate from "../truncate";
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { Icon, Table, Popup } from "semantic-ui-react";
 import Button from "react-bootstrap/Button";
-import { FiEdit3 } from "react-icons/fi";
 
 // import { useNavigate } from "react-router-dom";
 
@@ -104,7 +103,6 @@ export default function Course(props) {
   ];
 
   const job = [
-
     "Jr_Product_Designer",
     "Jr_Prod_Designer_Developer",
     "Innovation_Design_Thinking",
@@ -143,7 +141,7 @@ export default function Course(props) {
         </div>
         <Table className="coursestable" celled collapsing>
           <Table.Header>
-          <Table.Row>
+            <Table.Row>
               <Table.HeaderCell textAlign="center">ITI NAME</Table.HeaderCell>
               {headerTitles.map((title, index) => (
                 <Table.HeaderCell key={index} className="th2">
@@ -164,81 +162,23 @@ export default function Course(props) {
                   {job.map((jobTitle) => (
                     <Table.Cell key={jobTitle} textAlign="center" selectable>
                       {data[jobTitle].changeStatus === 1 ? (
-                        <>
-                          <Popup
-                            hoverable
-                            trigger={
-                              <div className="inside">
-                                {renderIcon(data[jobTitle].changeStatus)}
-                              </div>
-                            }
-                            content={
-                              <div style={{ width: "15vw", height: "10vh" }}>
-                                <p>
-                                  Start Date:{" "}
-                                  {formatDate(data[jobTitle].startDate)}
-                                </p>
-                                <p>
-                                  End Date: {formatDate(data[jobTitle].endDate)}
-                                </p>
-                              </div>
-                            }
-                          />
-                          <Popup
-                            trigger={
-                              <div className="inside">
-                                <FiEdit3 />
-                              </div>
-                            }
-                            content={
-                              <div style={{ width: "22vw", height: "25vh" }}>
-                                <form
-                                  className="noticeform"
-                                  onSubmit={(e) =>
-                                    handleSubmit(e, jobTitle, data.customId)
-                                  }
-                                >
-                                  <div className="titlearea">
-                                    Start date:
-                                    <input
-                                      className="inputtitle"
-                                      type="date"
-                                      value={startDate}
-                                      onChange={(e) =>
-                                        setStartDate(e.target.value)
-                                      }
-                                      required
-                                    />
-                                  </div>
-
-                                  <div className="titlearea">
-                                    End Date:
-                                    <input
-                                      className="inputtitle"
-                                      type="date"
-                                      value={endDate}
-                                      onChange={(e) =>
-                                        setEndDate(e.target.value)
-                                      }
-                                      required
-                                    />
-                                  </div>
-
-                                  <div className="buttonnotice">
-                                    <Button
-                                      className="addnotice"
-                                      type="submit"
-                                      variant="primary"
-                                    >
-                                      Click to Add
-                                    </Button>
-                                  </div>
-                                </form>
-                              </div>
-                            }
-                            on="click"
-                          />
-                        </>
+                        <Popup
+                          hoverable
+                          trigger={
+                            <div>{renderIcon(data[jobTitle].changeStatus)}</div>
+                          }
+                          content={
+                            <div style={{ width: "15vw", height: "10vh" }}>
+                              <p>
+                                Start Date:{" "}
+                                {formatDate(data[jobTitle].startDate)}
+                              </p>
+                              <p>
+                                End Date: {formatDate(data[jobTitle].endDate)}
+                              </p>
+                            </div>
+                          }
+                        />
                       ) : (
                         <Popup
                           trigger={
@@ -302,35 +242,3 @@ export default function Course(props) {
     </>
   );
 }
-
-// <Table.Row>
-//           <Table.Cell>Cell1</Table.Cell>
-//           <Table.Cell
-//             onMouseEnter={() => setHovered(true)}
-//             onMouseLeave={() => setHovered(false)}
-//             style={{ position: "absolute"  }}
-//           >
-//             Cellone
-//             {hovered && (
-//               <div className="dropdown-box-1">
-//                 <p>hello world</p>
-//                 <p>
-//                   Lorem ipsum dolor
-//                 </p>
-//               </div>
-//             )}
-//           </Table.Cell>
-//           <Table.Cell>
-//            cell
-//           </Table.Cell>
-//         </Table.Row>
-//         <Table.Row>
-//           <Table.Cell>Cell</Table.Cell>
-//           <Table.Cell>Cell</Table.Cell>
-//           <Table.Cell>Cell</Table.Cell>
-//         </Table.Row>
-//         <Table.Row>
-//           <Table.Cell>Cell</Table.Cell>
-//           <Table.Cell>Cell</Table.Cell>
-//           <Table.Cell>Cell</Table.Cell>
-//         </Table.Row>
