+++ src/tables/truncate.js
